bend/requirements.txt
# --- Core Framework ---
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
pydantic-settings==2.1.0

# --- Database (Postgres) ---
sqlalchemy==2.0.25
sqlmodel==0.0.14
# asyncpg is required for the "postgresql+asyncpg" URL in K3s
asyncpg==0.29.0
greenlet==3.0.3

# --- Storage (MinIO/S3) ---
# Official AWS SDK used to talk to MinIO
boto3==1.34.34
# For handling file uploads in FastAPI
python-multipart==0.0.6

# --- Security & Auth ---
# For JWT Token generation/validation
python-jose[cryptography]==3.3.0
# For hashing passwords
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- Production Utilities ---
# For CORS and cross-namespace communication
httpx==0.26.0
python-dotenv==1.0.1
# minIO
minio==7.2.0

# --- Rate Limiting ---
slowapi==0.1.9

# --- File Validation ---
python-magic==0.4.27

# --- Testing ---
pytest==8.0.0
pytest-asyncio==0.23.3
aiosqlite==0.19.0
